refactor: put most of the code in the workflows module instead of the pages directory
