+++ packages/twenty-front/src/modules/workflow/components/nodes/common/StyledTargetHandle.tsx
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-
 import { Handle } from '@xyflow/react';
 
 export const StyledTargetHandle = styled(Handle)`
